task/src/main/java/dev/elison/task/controller/TaskController.java
package dev.elison.task.controller;

import dev.elison.task.common.Task;
import dev.elison.task.service.TaskService;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/task")
public class TaskController {

    private final TaskService taskService;

    public TaskController(TaskService taskService) {
        this.taskService = taskService;
    }

    @GetMapping("/")
    public Iterable<Task> getTasks() {
        return taskService.getTasks();
    }

    @GetMapping("/{id}")
    public Task getTask(@PathVariable Long id) {
        return taskService.getTask(id);
    }

    @PostMapping("/")
    public void addTask(@RequestBody Task task) {
        taskService.saveTask(task);
    }

    @PutMapping("/")
    public void updateTask(@RequestBody Task task) {
        taskService.saveTask(task);
    }

    @DeleteMapping("/{id}")
    public void deleteTask(@PathVariable Long id) {
        taskService.deleteTask(id);
    }
}
